Put the georadius reply management in Reply
